feat: add JS module, local storage for toggle
