+++ homeworks/antonina-vechorko/05-css-advanced/index.html
@@ -58,7 +58,7 @@
                         <img src="assets/icons/arrow-btn.svg" alt="Arrow">
                     </span>
             </button>
-            <button class="person__btn person__btn--contact">Contact Me<a href="mailto:example@example.com"></a></button>
+            <button class="person__btn person__btn--contact"><a href="mailto:example@example.com">Contact Me</a></button>
         </div>
     </section>
 </header>
@@ -191,6 +191,6 @@
 <button class="scroll-to-top-btn" title="Scroll-Up">
     <img src="assets/icons/scroll-to-top.svg" alt="Scroll-Up">
 </button>
-<script src="index.js"></script>
+<script type="module" src="script/index.js"></script>
 </body>
 </html>
