+++ homeworks/antonina-vechorko/05-css-advanced/script/index.js
@@ -1,0 +1,4 @@
+import './dark-mode.js';
+import './toggle-switch.js';
+import './scroll-up.js';
+import './animation.js';
