+++ homeworks/antonina-vechorko/05-css-advanced/script/scroll-up.js
@@ -1,0 +1,28 @@
+const scrollToTopBtn = document.querySelector('.scroll-to-top-btn');
+
+document.addEventListener('DOMContentLoaded', () => {
+    scrollToTopBtn.addEventListener('click', () => {
+        window.scrollTo({
+            top: 0,
+            behavior: 'smooth'
+        });
+    });
+
+    const handleScroll = () => {
+        const scrollPosition = window.pageYOffset;
+        const viewportWidth = window.innerWidth;
+
+        if (viewportWidth <= 767 && scrollPosition > 500) {
+            // Show the button
+            scrollToTopBtn.style.display = 'block';
+        } else {
+            // Hide the button
+            scrollToTopBtn.style.display = 'none';
+        }
+    };
+
+    window.addEventListener('scroll', handleScroll);
+    window.addEventListener('resize', handleScroll);
+
+    handleScroll();
+});
